Completed organizing rover_new_op_pg
