+++ lib/ui/screens/rover_metrics_page.dart
@@ -1,10 +1,7 @@
 import 'dart:convert';
 
 import 'package:flutter/material.dart';
-import 'package:test/main.dart';
-import 'package:test/models/rover_health_type.dart';
 import 'package:test/models/rover_metrics.dart';
-import 'package:test/ui/screens/rover_operation_page.dart';
 import 'package:http/http.dart' as http;
 import 'package:test/services/mirv_api.dart';
